requirements.txt
Flask
google-generativeai
google-ai-generativelanguage